Drastically simplify focus zone code
